requirements.txt
Routes
eventlet
greenlet
six
webob
jinja2
PyYAML